ref: finder page rename and change cell element to PlayInfoView
